+++ DirectShowFilters/MPAudioRenderer/SoundTouch/source/RateTransposer.h
@@ -67,7 +67,7 @@ protected:
     /// Anti-alias filter object
     AAFilter *pAAFilter;
 
-    float fRate;
+    double fRate;
 
     int numChannels;
 
@@ -138,7 +138,7 @@ public:
 
     /// Sets new target rate. Normal rate = 1.0, smaller values represent slower 
     /// rate, larger faster rates.
-    virtual void setRate(float newRate);
+    virtual void setRate(double newRate);
 
     /// Sets the number of channels, 1 = mono, 2 = stereo
     void setChannels(int channels);
