+++ DirectShowFilters/MPAudioRenderer/SoundTouch/source/TDStretch.cpp
@@ -559,7 +559,7 @@ void TDStretch::calcSeqParameters()
 
 // Sets new target tempo. Normal tempo = 'SCALE', smaller values represent slower 
 // tempo, larger faster tempo.
-void TDStretch::setTempo(float newTempo)
+void TDStretch::setTempo(double newTempo)
 {
     int intskip;
 
@@ -942,14 +942,14 @@ long TDStretch::calcCrossCorrStereo(const short *mixingPos, const short *compare
 void TDStretch::precalcCorrReferenceStereo()
 {
     int i, cnt2;
-    float temp;
+    double temp;
 
     for (i=0 ; i < (int)overlapLength ;i ++) 
     {
-        temp = (float)i * (float)(overlapLength - i);
+        temp = (double)i * (double)(overlapLength - i);
         cnt2 = i * 2;
-        pRefMidBuffer[cnt2] = (float)(pMidBuffer[cnt2] * temp);
-        pRefMidBuffer[cnt2 + 1] = (float)(pMidBuffer[cnt2 + 1] * temp);
+        pRefMidBuffer[cnt2] = (double)(pMidBuffer[cnt2] * temp);
+        pRefMidBuffer[cnt2 + 1] = (double)(pMidBuffer[cnt2 + 1] * temp);
     }
 }
 
@@ -959,31 +959,31 @@ void TDStretch::precalcCorrReferenceStereo()
 void TDStretch::precalcCorrReferenceMono()
 {
     int i;
-    float temp;
+    double temp;
 
     for (i=0 ; i < (int)overlapLength ;i ++) 
     {
-        temp = (float)i * (float)(overlapLength - i);
-        pRefMidBuffer[i] = (float)(pMidBuffer[i] * temp);
+        temp = (double)i * (double)(overlapLength - i);
+        pRefMidBuffer[i] = (double)(pMidBuffer[i] * temp);
     }
 }
 
 
 // Overlaps samples in 'midBuffer' with the samples in 'pInput'
-void TDStretch::overlapStereo(float *pOutput, const float *pInput) const
+void TDStretch::overlapStereo(double *pOutput, const double *pInput) const
 {
     int i;
     int cnt2;
-    float fTemp;
-    float fScale;
-    float fi;
+    double fTemp;
+    double fScale;
+    double fi;
 
-    fScale = 1.0f / (float)overlapLength;
+    fScale = 1.0f / (double)overlapLength;
 
     for (i = 0; i < (int)overlapLength ; i ++) 
     {
-        fTemp = (float)(overlapLength - i) * fScale;
-        fi = (float)i * fScale;
+        fTemp = (double)(overlapLength - i) * fScale;
+        fi = (double)i * fScale;
         cnt2 = 2 * i;
         pOutput[cnt2 + 0] = pInput[cnt2 + 0] * fi + pMidBuffer[cnt2 + 0] * fTemp;
         pOutput[cnt2 + 1] = pInput[cnt2 + 1] * fi + pMidBuffer[cnt2 + 1] * fTemp;
@@ -1008,7 +1008,7 @@ void TDStretch::calculateOverlapLength(int overlapInMsec)
 
 
 
-double TDStretch::calcCrossCorrMono(const float *mixingPos, const float *compare) const
+double TDStretch::calcCrossCorrMono(const double *mixingPos, const double *compare) const
 {
     double corr;
     double norm;
@@ -1026,7 +1026,7 @@ double TDStretch::calcCrossCorrMono(const float *mixingPos, const float *compare
 }
 
 
-double TDStretch::calcCrossCorrStereo(const float *mixingPos, const float *compare) const
+double TDStretch::calcCrossCorrStereo(const double *mixingPos, const double *compare) const
 {
     double corr;
     double norm;
