+++ DirectShowFilters/MPAudioRenderer/source/MultiSoundTouch.cpp
@@ -165,9 +165,9 @@ void CMultiSoundTouch::StopResamplingThread()
 DEFINE_STREAM_FUNC(setRate, double, newRate)
 DEFINE_STREAM_FUNC(setRateChange, double, newRate)
 DEFINE_STREAM_FUNC(setTempoChange, double, newTempo)
-DEFINE_STREAM_FUNC(setPitchOctaves, float, newPitch)
+DEFINE_STREAM_FUNC(setPitchOctaves, double, newPitch)
 DEFINE_STREAM_FUNC(setPitchSemiTones, int, newPitch)
-DEFINE_STREAM_FUNC(setPitchSemiTones, float, newPitch)
+DEFINE_STREAM_FUNC(setPitchSemiTones, double, newPitch)
 DEFINE_STREAM_FUNC(setSampleRate, uint, srate)
 
 // clear requires a specific handling since we need to be able to use the CAutoLock
